requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.7.0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
cppimport==21.3.7
discord.py==1.7.1
gitdb==4.0.7
GitPython==3.1.14
idna==2.10
Mako==1.1.4
MarkupSafe==1.1.1
multidict==5.1.0
Pillow==8.2.0
praw==7.2.0
prawcore==2.0.0
pybind11==2.6.2
pyodbc==4.0.30
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.15.0
smmap==4.0.0
typing-extensions==3.7.4.3
tzlocal==2.1
update-checker==0.18.0
urllib3==1.26.4
websocket-client==0.58.0
yarl==1.6.3
